Supplant but keep old
